modify use-aliyun-mirror.sh

+++ use-aliyun-mirror.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
-mv /etc/yum.repos.d/CentOS-Base.repo /etc/yum.repos.d/CentOS-Base.repo.backup
-wget -O /etc/yum.repos.d/CentOS-Base.repo http://mirrors.aliyun.com/repo/Centos-7.repo
+if [ -e /etc/yum.repos.d/CentOS-Base.repo.backup ]; then
+    echo 'already exist backup file, skip download'
+else
+    mv /etc/yum.repos.d/CentOS-Base.repo /etc/yum.repos.d/CentOS-Base.repo.backup
+    wget -O /etc/yum.repos.d/CentOS-Base.repo http://mirrors.aliyun.com/repo/Centos-7.repo
+fi
 sed -i -e '/mirrors.cloud.aliyuncs.com/d' -e '/mirrors.aliyuncs.com/d' /etc/yum.repos.d/CentOS-Base.repo
 yum makecache
 yum upgrade -y
